divide test files
