+++ common/perception_utils/test/src/test_object_classification.hpp
@@ -1,0 +1,63 @@
+// Copyright 2020 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "perception_utils/object_classification.hpp"
+
+#include <gtest/gtest.h>
+
+namespace
+{
+autoware_auto_perception_msgs::msg::ObjectClassification createObjectClassification(
+  const std::uint8_t label, const double probability)
+{
+  autoware_auto_perception_msgs::msg::ObjectClassification classification;
+  classification.label = label;
+  classification.probability = probability;
+
+  return classification;
+}
+
+}  // namespace
+
+TEST(perception_utils, test_getHighestProbLabel)
+{
+  using autoware_auto_perception_msgs::msg::ObjectClassification;
+  using perception_utils::getHighestProbLabel;
+
+  {  // empty
+    std::vector<autoware_auto_perception_msgs::msg::ObjectClassification> classification;
+    std::uint8_t label = getHighestProbLabel(classification);
+    EXPECT_EQ(label, ObjectClassification::UNKNOWN);
+  }
+
+  {  // normal case
+    std::vector<autoware_auto_perception_msgs::msg::ObjectClassification> classification;
+    classification.push_back(createObjectClassification(ObjectClassification::CAR, 0.5));
+    classification.push_back(createObjectClassification(ObjectClassification::TRUCK, 0.8));
+    classification.push_back(createObjectClassification(ObjectClassification::BUS, 0.7));
+
+    std::uint8_t label = getHighestProbLabel(classification);
+    EXPECT_EQ(label, ObjectClassification::TRUCK);
+  }
+
+  {  // labels with the same probability
+    std::vector<autoware_auto_perception_msgs::msg::ObjectClassification> classification;
+    classification.push_back(createObjectClassification(ObjectClassification::CAR, 0.8));
+    classification.push_back(createObjectClassification(ObjectClassification::TRUCK, 0.8));
+    classification.push_back(createObjectClassification(ObjectClassification::BUS, 0.7));
+
+    std::uint8_t label = getHighestProbLabel(classification);
+    EXPECT_EQ(label, ObjectClassification::CAR);
+  }
+}
